requirements.txt
astunparse==1.6.3
bcrypt==3.1.7
blinker==1.4
cachelib==0.1.1
cachetools==4.2.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==2.8
cycler==0.10.0
dbus-python==1.2.16
decorator==4.4.2
detecto==1.2.0
distro==1.4.0
dotmap==1.3.23
einops==0.2.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
Flask==1.1.2
Flask-Session==0.3.2
flatbuffers==1.12
future==0.18.2
gast==0.3.3
gdown==3.11.0
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
httplib2==0.14.0
idna==2.8
imageio==2.9.0
imgaug==0.4.0
imutils==0.5.3
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
lmdb==1.0.0
lockfile==0.12.2
lxml==4.6.1
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.3
MarkupSafe==1.1.0
matplotlib==3.1.2
mccabe==0.6.1
monotonic==1.5
netifaces==0.10.4
networkx==2.5
numpy==1.19.4
oauthlib==3.1.0
olefile==0.46
opencv-python==4.4.0.46
opt-einsum==3.3.0
pandas==1.1.4
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
prefetch-generator==1.0.1
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
pylint==2.6.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
PyQt5==5.15.2
PyQt5-sip==12.8.1
pyRFC3339==1.1
PySocks==1.7.1
python-dateutil==2.7.3
pytz==2019.3
PyWavelets==1.1.1
pyxdg==0.26
PyYAML==5.3.1
redis==3.5.3
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.6
scikit-image==0.18.0
scipy==1.5.4
SecretStorage==2.3.1
Shapely==1.7.1
simplejson==3.16.0
six==1.15.0
dbus-python==1.2.16
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
torch-1.7.1-cp38-cp38-manylinux1_x86_64.whl
torchvision==0.8.2
tqdm==4.54.1
vietocr==0.3.5
